CypherRAG/diffbot_final_app/pyproject.toml
[tool.poetry]
name = "diffbot_demo_app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
openai = "^1.29.0"
langchain = "^0.1.20"
python-dotenv = "^1.0.1"
neo4j = "^5.20.0"
langchain-community = "^0.0.38"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
